config/settings.py
import os
from pathlib import Path

import environ

BASE_DIR = Path(__file__).resolve().parent.parent

env = environ.Env(
    DEBUG=(bool, False),
    ALLOWED_HOSTS=(list, ["*"]),
    REDIS_URL=(str, "redis://redis:6379/0"),
    # When False and DEBUG is False, only literal ALLOWED_HOSTS from .env are used.
    ALLOW_NGROK_SUBDOMAINS=(bool, False),
    # Cloud Run / load balancers terminate TLS; set True in production on GCP.
    TRUST_X_FORWARDED_PROTO=(bool, False),
)

env.read_env(os.path.join(BASE_DIR, ".env"))

SECRET_KEY = env("SECRET_KEY", default="change-me-in-production")

DEBUG = env("DEBUG")


def _normalized_allowed_hosts(raw: list | str | tuple) -> list[str]:
    if isinstance(raw, str):
        parts = [p.strip() for p in raw.split(",")]
    else:
        parts = [str(p).strip() for p in raw]
    return [p for p in parts if p]


# Ngrok free URLs change every run; Django accepts any *.suffix when suffix is listed
# with a leading dot (e.g. .ngrok-free.app). Trailing commas in .env can inject "" — drop them.
_NGROK_WILDCARD_SUFFIXES = (
    ".ngrok-free.app",
    ".ngrok-free.dev",
    ".ngrok.io",
    ".ngrok.app",
)

_raw_allowed = env("ALLOWED_HOSTS")
_hosts = _normalized_allowed_hosts(_raw_allowed)

if "*" in _hosts:
    ALLOWED_HOSTS = ["*"]
elif DEBUG or env("ALLOW_NGROK_SUBDOMAINS"):
    ALLOWED_HOSTS = list(dict.fromkeys([*_hosts, *_NGROK_WILDCARD_SUFFIXES]))
else:
    ALLOWED_HOSTS = _hosts

if env("TRUST_X_FORWARDED_PROTO"):
    SECURE_PROXY_SSL_HEADER = ("HTTP_X_FORWARDED_PROTO", "https")

# ---------------------------------------------------------------------------
# Application definition
# ---------------------------------------------------------------------------

INSTALLED_APPS = [
    "daphne",
    "django.contrib.admin",
    "django.contrib.auth",
    "django.contrib.contenttypes",
    "django.contrib.sessions",
    "django.contrib.messages",
    "django.contrib.staticfiles",
    # Third-party
    "rest_framework",
    "corsheaders",
    "channels",
    "rest_framework_simplejwt",
    # Local
    "core",
    "channels_app",
    "escalation",
    "teams",
]

MIDDLEWARE = [
    "corsheaders.middleware.CorsMiddleware",
    "django.middleware.security.SecurityMiddleware",
    "django.contrib.sessions.middleware.SessionMiddleware",
    "django.middleware.common.CommonMiddleware",
    "django.middleware.csrf.CsrfViewMiddleware",
    "django.contrib.auth.middleware.AuthenticationMiddleware",
    "django.contrib.messages.middleware.MessageMiddleware",
    "django.middleware.clickjacking.XFrameOptionsMiddleware",
    "teams.middleware.TeamMiddleware",
]

ROOT_URLCONF = "config.urls"

TEMPLATES = [
    {
        "BACKEND": "django.template.backends.django.DjangoTemplates",
        "DIRS": [BASE_DIR / "templates"],
        "APP_DIRS": True,
        "OPTIONS": {
            "context_processors": [
                "django.template.context_processors.debug",
                "django.template.context_processors.request",
                "django.contrib.auth.context_processors.auth",
                "django.contrib.messages.context_processors.messages",
            ],
        },
    },
]

ASGI_APPLICATION = "config.asgi.application"

WSGI_APPLICATION = "config.wsgi.application"

# ---------------------------------------------------------------------------
# Database
# ---------------------------------------------------------------------------

DATABASES = {
    "default": env.db("DATABASE_URL", default="sqlite:///db.sqlite3"),
}

# ---------------------------------------------------------------------------
# Password validation
# ---------------------------------------------------------------------------

AUTH_PASSWORD_VALIDATORS = [
    {"NAME": "django.contrib.auth.password_validation.UserAttributeSimilarityValidator"},
    {"NAME": "django.contrib.auth.password_validation.MinimumLengthValidator"},
    {"NAME": "django.contrib.auth.password_validation.CommonPasswordValidator"},
    {"NAME": "django.contrib.auth.password_validation.NumericPasswordValidator"},
]

# ---------------------------------------------------------------------------
# Internationalization
# ---------------------------------------------------------------------------

LANGUAGE_CODE = "en-us"
TIME_ZONE = "UTC"
USE_I18N = True
USE_TZ = True

# ---------------------------------------------------------------------------
# Static files
# ---------------------------------------------------------------------------

STATIC_URL = "static/"
STATIC_ROOT = BASE_DIR / "staticfiles"
STATICFILES_DIRS = [BASE_DIR / "static"] if (BASE_DIR / "static").exists() else []

MEDIA_URL = "/api/media/"
MEDIA_ROOT = BASE_DIR / "media"

# ---------------------------------------------------------------------------
# Default primary key field type
# ---------------------------------------------------------------------------

DEFAULT_AUTO_FIELD = "django.db.models.BigAutoField"

# ---------------------------------------------------------------------------
# Django REST Framework
# ---------------------------------------------------------------------------

REST_FRAMEWORK = {
    "DEFAULT_RENDERER_CLASSES": [
        "rest_framework.renderers.JSONRenderer",
    ],
    "DEFAULT_AUTHENTICATION_CLASSES": [
        "rest_framework_simplejwt.authentication.JWTAuthentication",
        "rest_framework.authentication.SessionAuthentication",
    ],
    "DEFAULT_PAGINATION_CLASS": "rest_framework.pagination.PageNumberPagination",
    "PAGE_SIZE": 20,
}

# ---------------------------------------------------------------------------
# Simple JWT
# ---------------------------------------------------------------------------

from datetime import timedelta  # noqa: E402

SIMPLE_JWT = {
    "ACCESS_TOKEN_LIFETIME": timedelta(hours=1),
    "REFRESH_TOKEN_LIFETIME": timedelta(days=7),
    "ROTATE_REFRESH_TOKENS": True,
    "BLACKLIST_AFTER_ROTATION": False,
}

# ---------------------------------------------------------------------------
# CORS
# ---------------------------------------------------------------------------

CORS_ALLOW_ALL_ORIGINS = True  # For development only

# ---------------------------------------------------------------------------
# Django Channels / Redis
# ---------------------------------------------------------------------------

CHANNEL_LAYERS = {
    "default": {
        "BACKEND": "channels_redis.core.RedisChannelLayer",
        "CONFIG": {
            "hosts": [env("REDIS_URL")],
        },
    },
}

# ---------------------------------------------------------------------------
# WhatsApp
# ---------------------------------------------------------------------------

WHATSAPP_ACCESS_TOKEN = env("WHATSAPP_ACCESS_TOKEN", default="")
WHATSAPP_PHONE_NUMBER_ID = env("WHATSAPP_PHONE_NUMBER_ID", default="")
WHATSAPP_VERIFY_TOKEN = env("WHATSAPP_VERIFY_TOKEN", default="")

# ---------------------------------------------------------------------------
# Anthropic
# ---------------------------------------------------------------------------

ANTHROPIC_API_KEY = env("ANTHROPIC_API_KEY", default="")
CLAUDE_HAIKU_MODEL = env("CLAUDE_HAIKU_MODEL", default="claude-haiku-4-5-20251001")
CLAUDE_SONNET_MODEL = env("CLAUDE_SONNET_MODEL", default="claude-sonnet-4-5-20250929")

# ---------------------------------------------------------------------------
# OpenAI (embeddings)
# ---------------------------------------------------------------------------

OPENAI_API_KEY = env("OPENAI_API_KEY", default="")

# ---------------------------------------------------------------------------
# Telegram
# ---------------------------------------------------------------------------

TELEGRAM_BOT_TOKEN = env("TELEGRAM_BOT_TOKEN", default="")

# ---------------------------------------------------------------------------
# Google / Gmail
# ---------------------------------------------------------------------------

# ---------------------------------------------------------------------------
# Messenger / Instagram
# ---------------------------------------------------------------------------

MESSENGER_PAGE_ACCESS_TOKEN = env("MESSENGER_PAGE_ACCESS_TOKEN", default="")
MESSENGER_VERIFY_TOKEN = env("MESSENGER_VERIFY_TOKEN", default="")

# ---------------------------------------------------------------------------
# Voice (Twilio + optional ElevenLabs)
# ---------------------------------------------------------------------------

PUBLIC_BASE_URL = env("PUBLIC_BASE_URL", default="")
TWILIO_ACCOUNT_SID = env("TWILIO_ACCOUNT_SID", default="")
TWILIO_AUTH_TOKEN = env("TWILIO_AUTH_TOKEN", default="")
ELEVENLABS_API_KEY = env("ELEVENLABS_API_KEY", default="")
ELEVENLABS_VOICE_ID = env("ELEVENLABS_VOICE_ID", default="")
ELEVENLABS_MODEL_ID = env("ELEVENLABS_MODEL_ID", default="eleven_turbo_v2_5")
VOICE_ESCALATION_FORWARD_NUMBER = env("VOICE_ESCALATION_FORWARD_NUMBER", default="")

GOOGLE_CREDENTIALS_PATH = env("GOOGLE_CREDENTIALS_PATH", default="")
GMAIL_WATCH_ADDRESS = env("GMAIL_WATCH_ADDRESS", default="")
GOOGLE_CLIENT_ID = env("GOOGLE_CLIENT_ID", default="")
GOOGLE_CLIENT_SECRET = env("GOOGLE_CLIENT_SECRET", default="")
