requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
psycopg2-binary>=2.9,<3.0
pgvector>=0.3,<1.0
anthropic>=0.40,<1.0
requests>=2.31,<3.0
google-auth>=2.29,<3.0
google-auth-oauthlib>=1.2,<2.0
google-auth-httplib2>=0.2,<1.0
google-api-python-client>=2.120,<3.0
gunicorn>=22.0,<23.0
openai>=1.30,<2.0
djangorestframework-simplejwt>=5.3,<6.0
pypdf>=4.0,<5.0
twilio>=9.0,<10.0
